third_party/blink/web_tests/printing/pseudo-class-outside-page-expected.txt

Test case for SHOULD NEVER BE REACHED assertion loading forbes.com. Paged media related pseudo-classes should be ignored outside of @page.

You should see a green box for Test 1, "Test n: PASS" for the rest.

Test 2: PASS
Test 3: PASS
Test 4: PASS
Test 5: PASS
Test 6: PASS
Test 7: PASS
Test 8: PASS
Test 9: PASS
Test 10: PASS
Test 11: PASS
Test 12: PASS
Test 13: PASS
PASS
